// client/src/components/api/MenuScheme.tsx
interface MenuScheme 
      {
        _id:string,
        rname: string;
        imgdata: string;
        address: string;
        delimg: string;
        somedata: string;
        price: number;
        rating: string;
        arrimg: string;
        qnty: number;
      }
    
export default MenuScheme;